refactor(accounting): move DTO types from presentation to domain [DTO cluster]
First paydown of the frozen hexagonal debt. The journal/ledger/accounts
DTO types lived in presentation/dto/*.types but described the accounting
domain, so domain/ports and application/ reached UP into presentation to
use them — R1 (domain depends on presentation) and R2 (application depends
on presentation), the inverted-arrow violations that are structurally the
worst kind.

Move the 3 files to modules/accounting/domain/ (flat, matching the
audit/documents/rag/tags convention of domain/*.types.ts at the root).
Zero runtime exports — pure interface/type files, so the move is
runtime- and bundling-safe. 30 specifiers rewritten across domain/ports,
application, infrastructure and presentation, plus the 4 external
consumers (account-balances, payment, two components/accounting files)
and the type-only re-exports in presentation/server.ts.

Ratchet baseline: 138 → 131 (net -7).
- Removed 10 resolved entries: R1×3, R2×7. Accounting's R1 class drops
  from 13 to 10.
- Added 3 entries: domain/{accounts,journal,ledger}.types.ts:R5. This is
  not new careless debt — it is the sanctioned [DTO]→[PRISMA] handoff. The
  R5 allowTypeImports carve-out is granted to presentation/ only, so
  type-only Prisma imports that were legal in presentation surface as R5
  the moment the files live in domain. Seven sibling domain type files
  (equity-statement, financial-statements, trial-balance, worksheet, ...)
  are already baselined with :R5 from the identical pattern, and the
  baseline legend itself prescribes this move. The carve-out was NOT
  extended to domain: the domain must stay Prisma-agnostic even in types,
  so that R5 is real [PRISMA]-cluster debt, closed later by domain-local
  types + mapping at the infra boundary.

Header prose updated 138 → 131, with the 138 kept as the documented
starting point. A ratchet that misreports its own frozen count is the rot
it exists to prevent.

Three shape tests pinned the OLD presentation/ path and were updated to
domain/ with a NOTE — two of them wrote the path as an ESCAPED regex
(presentation\/dto), which evaded the initial grep and only surfaced under
the full suite. Sweep both escaped and unescaped forms when relocating
files that shape tests read via readFileSync+regex.

tsc clean. 783 files / 7303 tests, 3 pre-existing failures in
payment-form-shortcut.test.tsx. Lint no-restricted-imports: 131.
