+++ modules/accounting/domain/journal.types.ts
